requirements.txt
APScheduler==3.6.3
asgiref==3.5.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
importlib-metadata==4.11.2
Markdown==3.3.6
Pillow==9.0.1
pyaes==1.6.1
pyasn1==0.4.8
PyJWT==2.3.0
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
telebot==0.0.4
Telethon==1.24.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
zipp==3.7.0
